backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pydantic==2.12.5
python-telegram-bot==22.6
python-dotenv==1.2.2
supabase==2.28.2
pdfplumber==0.11.9
google-cloud-storage==3.9.0
google-cloud-aiplatform==1.141.0
vertexai==1.43.0
google-genai==1.67.0
python-multipart==0.0.22
pytz==2026.1.post1
